==> notebooks/05-model-inference.ipynb <==
# Model Inference Notebook

Use the trained model for prediction on new data

# Steps

1. Split the data into training and test data set
1. Clean the data (transform null values)
1. Scale necessary attributes (normalization, standardization)
1. Save transformed data for model training


# Class to load model and run prediction

import pandas as pd
import logging
import joblib
biomarkers = None
pipeline = None
model = None

def transformAndPredict(raw, pipeline, model, dropNonBioMarkers = True) :
    print ("predict")
    # dropping non-bio markers
    if dropNonBioMarkers :
        dropped = raw.drop(
        ["Age", "Unit1", "Unit2", "HospAdmTime", "ICULOS", "Gender", "Bilirubin_direct", "TroponinI", "isSepsis"],
          axis=1)
    else :
        dropped = raw

    # perform inference and predict
    transformed = pipeline.transform(dropped)
    prediction = model.predict(transformed)
    results = pd.DataFrame(prediction)
    results.columns = ["issepsis"]

    return results


# Load data and run prediction

Load semi-colon separated data from disk

import sys
sys.path.append("../lib")
from getConfig import *
config = getConfig("../")

 
#Load the data to run prediction on, model and pipeline
inference_data = pd.read_csv(config.inference_path, sep=",")
pipeline = joblib.load(config.traintest_path + "/pipeline.pkl")
model = joblib.load(config.tuned_path + "rfc_model.pkl")

        
# Run prediction
results = transformAndPredict(inference_data, pipeline, model, dropNonBioMarkers = True )

# print results
for index, row in results.iterrows():
    issepsis = row["issepsis"]
print(results)

fp = open(config.experiments_path + "prediction_results.txt", "w")
fp.write(str(results))
fp.close()
